create custom-state-history-animation module for generating state history animations
